Refactor

+++ src/main/java/org/ademun/timetableapi/repository/GroupRepository.java
@@ -1,9 +1,9 @@
 package org.ademun.timetableapi.repository;
 
-import org.ademun.timetableapi.model.Group;
+import org.ademun.timetableapi.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 }
